Move library view logic into fragment
